rename the modules (again!) to reflect their domain
Protocol -> Messages (containing message definitions and their json instances)
Messages -> Peer  (functions that talk to the peer)
