+++ src/Transit/Internal/Messages.hs
@@ -1,67 +1,141 @@
-{-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE DeriveGeneric #-}
+{-# LANGUAGE ScopedTypeVariables #-}
 module Transit.Internal.Messages
-  ( makeSenderHandshake
-  , makeReceiverHandshake
-  , makeSenderRecordKey
-  , makeReceiverRecordKey
-  )
-where
+  ( TransitMsg(..)
+  , Ability(..)
+  , AbilityV1(..)
+  , Hint(..)
+  , ConnectionHint(..)
+  , Ack(..)
+  , TransitAck(..)
+  ) where
 
 import Protolude
 
-import qualified Crypto.KDF.HKDF as HKDF
-import Crypto.Hash (SHA256(..))
-import qualified Crypto.Saltine.Internal.ByteSizes as ByteSizes
-import qualified Crypto.Saltine.Class as Saltine
-import qualified Crypto.Saltine.Core.SecretBox as SecretBox
-import Data.Hex (hex)
-import Data.Text (toLower)
-
-hkdf :: ByteString -> SecretBox.Key -> ByteString -> ByteString
-hkdf salt key purpose =
-  HKDF.expand (HKDF.extract salt (Saltine.encode key) :: HKDF.PRK SHA256) purpose keySize
-  where
-    keySize = ByteSizes.secretBoxKey
-
-data Purpose
-  = SenderHandshake
-  | ReceiverHandshake
-  | SenderRecord
-  | ReceiverRecord
-  deriving (Eq, Show)
-
-deriveKeyFromPurpose :: Purpose -> SecretBox.Key -> ByteString
-deriveKeyFromPurpose purpose key =
-  hkdf salt key (purposeStr purpose)
-  where
-    salt = "" :: ByteString
-    purposeStr :: Purpose -> ByteString
-    purposeStr SenderHandshake = "transit_sender"
-    purposeStr ReceiverHandshake = "transit_receiver"
-    purposeStr SenderRecord = "transit_record_sender_key"
-    purposeStr ReceiverRecord = "transit_record_receiver_key"
-
-makeSenderHandshake :: SecretBox.Key -> ByteString
-makeSenderHandshake key =
-  (toS @Text @ByteString "transit sender ") <> hexid <> (toS @Text @ByteString " ready\n\n")
-  where
-    subkey = deriveKeyFromPurpose SenderHandshake key
-    hexid = (toS (toLower (toS @ByteString @Text (hex subkey))))
+import Data.Aeson
+  ( FromJSON(..)
+  , ToJSON(..)
+  , genericToJSON
+  , genericParseJSON
+  , defaultOptions
+  , fieldLabelModifier
+  , constructorTagModifier
+  , sumEncoding
+  , SumEncoding(..)
+  , camelTo2
+  )
 
 
-makeReceiverHandshake :: SecretBox.Key -> ByteString
-makeReceiverHandshake key =
-  (toS @Text @ByteString "transit receiver ") <> hexid <> (toS @Text @ByteString " ready\n\n")
-  where
-    subkey = deriveKeyFromPurpose ReceiverHandshake key
-    hexid = (toS (toLower (toS @ByteString @Text (hex subkey))))
+data AbilityV1
+  = DirectTcpV1
+  | RelayV1
+  deriving (Eq, Show, Generic)
 
-makeSenderRecordKey :: SecretBox.Key -> SecretBox.Key
-makeSenderRecordKey key =
-  fromMaybe (panic "Could not encode to SecretBox key") $
-  Saltine.decode (deriveKeyFromPurpose SenderRecord key)
+instance ToJSON AbilityV1 where
+  toJSON = genericToJSON
+    defaultOptions { constructorTagModifier = camelTo2 '-'}
 
-makeReceiverRecordKey :: SecretBox.Key -> SecretBox.Key
-makeReceiverRecordKey key =
-  fromMaybe (panic "Could not encode to SecretBox key") $
-  Saltine.decode (deriveKeyFromPurpose ReceiverRecord key)
+instance FromJSON AbilityV1 where
+  parseJSON = genericParseJSON
+    defaultOptions { constructorTagModifier = camelTo2 '-'}
+
+data Hint = Hint { ctype :: AbilityV1
+                 , priority :: Double
+                 , hostname :: Text
+                 , port :: Word16 }
+          deriving (Eq, Show, Generic)
+
+instance ToJSON Hint where
+  toJSON = genericToJSON
+    defaultOptions { fieldLabelModifier =
+                       \name -> case name of
+                                  "ctype" -> "type"
+                                  _ -> name }
+
+instance FromJSON Hint where
+  parseJSON = genericParseJSON
+    defaultOptions { fieldLabelModifier =
+                       \name -> case name of
+                                  "ctype" -> "type"
+                                  _ -> name }
+
+data ConnectionHint
+  = Direct Hint
+  | Tor Hint
+  | Relay { rtype :: AbilityV1
+          , hints :: [Hint] }
+  deriving (Eq, Show, Generic)
+
+instance ToJSON ConnectionHint where
+  toJSON = genericToJSON
+    defaultOptions { sumEncoding = UntaggedValue
+                   , fieldLabelModifier =
+                       \name -> case name of
+                                  "rtype" -> "type"
+                                  _ -> name }
+instance FromJSON ConnectionHint where
+  parseJSON = genericParseJSON
+    defaultOptions { sumEncoding = UntaggedValue
+                   , fieldLabelModifier =
+                       \name -> case name of
+                                  "rtype" -> "type"
+                                  _ -> name }
+
+data Ack = FileAck Text
+         | MsgAck Text
+         deriving (Eq, Show, Generic)
+
+instance ToJSON Ack where
+  toJSON = genericToJSON
+    defaultOptions { sumEncoding = ObjectWithSingleField
+                   , constructorTagModifier = camelTo2 '_'}
+
+instance FromJSON Ack where
+  parseJSON = genericParseJSON
+    defaultOptions { sumEncoding = ObjectWithSingleField
+                   , constructorTagModifier = camelTo2 '_'}
+
+data Ability = Ability { atype :: AbilityV1 }
+  deriving (Eq, Show, Generic)
+
+instance ToJSON Ability where
+  toJSON = genericToJSON
+    defaultOptions { sumEncoding = UntaggedValue
+                   , fieldLabelModifier = const "type" }
+
+instance FromJSON Ability where
+  parseJSON = genericParseJSON
+    defaultOptions { sumEncoding = UntaggedValue
+                   , fieldLabelModifier = const "type" }
+
+data TransitMsg = Error Text
+                | Answer Ack
+                | Transit { abilitiesV1 :: [Ability]
+                          , hintsV1 :: [ConnectionHint] }
+                deriving (Eq, Show, Generic)
+
+instance ToJSON TransitMsg where
+  toJSON = genericToJSON
+    defaultOptions { sumEncoding = ObjectWithSingleField
+                   , constructorTagModifier = camelTo2 '-'
+                   , fieldLabelModifier = camelTo2 '-' }
+instance FromJSON TransitMsg where
+  parseJSON = genericParseJSON
+    defaultOptions { sumEncoding = ObjectWithSingleField
+                   , constructorTagModifier = camelTo2 '-'
+                   , fieldLabelModifier = camelTo2 '-'}
+
+data TransitAck
+  = TransitAck
+  { ack :: Text -- ^ "ack" is "ok" implies a successful transfer
+  , sha256 :: Text } -- ^ expected sha256 sum of the transfered file
+  deriving (Eq, Show, Generic)
+
+instance ToJSON TransitAck where
+  toJSON = genericToJSON
+    defaultOptions { sumEncoding = UntaggedValue }
+
+instance FromJSON TransitAck where
+  parseJSON = genericParseJSON
+    defaultOptions { sumEncoding = UntaggedValue }
+
