+++ src/Transit/Internal/Peer.hs
@@ -1,0 +1,67 @@
+{-# LANGUAGE OverloadedStrings #-}
+module Transit.Internal.Peer
+  ( makeSenderHandshake
+  , makeReceiverHandshake
+  , makeSenderRecordKey
+  , makeReceiverRecordKey
+  )
+where
+
+import Protolude
+
+import qualified Crypto.KDF.HKDF as HKDF
+import Crypto.Hash (SHA256(..))
+import qualified Crypto.Saltine.Internal.ByteSizes as ByteSizes
+import qualified Crypto.Saltine.Class as Saltine
+import qualified Crypto.Saltine.Core.SecretBox as SecretBox
+import Data.Hex (hex)
+import Data.Text (toLower)
+
+hkdf :: ByteString -> SecretBox.Key -> ByteString -> ByteString
+hkdf salt key purpose =
+  HKDF.expand (HKDF.extract salt (Saltine.encode key) :: HKDF.PRK SHA256) purpose keySize
+  where
+    keySize = ByteSizes.secretBoxKey
+
+data Purpose
+  = SenderHandshake
+  | ReceiverHandshake
+  | SenderRecord
+  | ReceiverRecord
+  deriving (Eq, Show)
+
+deriveKeyFromPurpose :: Purpose -> SecretBox.Key -> ByteString
+deriveKeyFromPurpose purpose key =
+  hkdf salt key (purposeStr purpose)
+  where
+    salt = "" :: ByteString
+    purposeStr :: Purpose -> ByteString
+    purposeStr SenderHandshake = "transit_sender"
+    purposeStr ReceiverHandshake = "transit_receiver"
+    purposeStr SenderRecord = "transit_record_sender_key"
+    purposeStr ReceiverRecord = "transit_record_receiver_key"
+
+makeSenderHandshake :: SecretBox.Key -> ByteString
+makeSenderHandshake key =
+  (toS @Text @ByteString "transit sender ") <> hexid <> (toS @Text @ByteString " ready\n\n")
+  where
+    subkey = deriveKeyFromPurpose SenderHandshake key
+    hexid = (toS (toLower (toS @ByteString @Text (hex subkey))))
+
+
+makeReceiverHandshake :: SecretBox.Key -> ByteString
+makeReceiverHandshake key =
+  (toS @Text @ByteString "transit receiver ") <> hexid <> (toS @Text @ByteString " ready\n\n")
+  where
+    subkey = deriveKeyFromPurpose ReceiverHandshake key
+    hexid = (toS (toLower (toS @ByteString @Text (hex subkey))))
+
+makeSenderRecordKey :: SecretBox.Key -> SecretBox.Key
+makeSenderRecordKey key =
+  fromMaybe (panic "Could not encode to SecretBox key") $
+  Saltine.decode (deriveKeyFromPurpose SenderRecord key)
+
+makeReceiverRecordKey :: SecretBox.Key -> SecretBox.Key
+makeReceiverRecordKey key =
+  fromMaybe (panic "Could not encode to SecretBox key") $
+  Saltine.decode (deriveKeyFromPurpose ReceiverRecord key)
