+++ tests/Main.hs
@@ -2,10 +2,10 @@ module Main (main) where
 
 import Protolude
 
-import qualified FileTransferTests
+import qualified ProtocolTests
 import qualified MessagesTests
 
 main :: IO ()
 main = do
-  FileTransferTests.tests
+  ProtocolTests.tests
   MessagesTests.tests
